refactor: rename search to analytics
