+++ frontend-vue/src/components/analytics/activeFilter.ts
